requirements.txt
kivy
pycryptodome
opencv-python
numpy
